+++ doc/version-history.rst
@@ -8,6 +8,7 @@ v2.8.0
 ------
 
 * Communicate with two flow meters.
+* Save and load glycol/heaters setpoints.
 * Read flow meter identification.
 
 v2.7.0